Adapted client to the 0.0.4 server version that allows flexible payloads. + Removed Spring AMQP as it's an overkill

+++ src/main/java/com/pablintino/schedulerservice/client/models/SchedulerMetadata.java
@@ -7,5 +7,5 @@ import java.time.Instant;
 @RequiredArgsConstructor
 public class SchedulerMetadata {
   private final Instant triggerTime;
-  private final int triggerAttempt;
+  private final long triggerAttempt;
 }
